bot/requirements.txt
# Bot Telegram
python-telegram-bot==20.7
python-dotenv==1.0.0

# Base de données
sqlite3
sqlalchemy==2.0.23
alembic==1.13.1

# API et communication
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
requests==2.31.0

# Tests
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Linting et formatting
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Logging
loguru==0.7.2

# Utilitaires
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1

# Cache et performance
redis==5.0.1
aioredis==2.0.1

# Sécurité
cryptography==41.0.8
passlib==1.7.4

# Monitoring
prometheus-client==0.19.0
